requirements.txt
autocorrect~=2.6.1
aiml~=0.9.2
flask~=3.1.0




flask==2.0.1
flask-cors
neo4j
#py2neo==2021.2.3
python-aiml==0.9.3
autocorrect==2.6.1
#python-dotenv==0.19.0
python-dotenv==0.21.0

#pyswip>=0.2.10
#regex>=2023.10.3
pyswip~=0.3.2
mysql-connector-python~=9.3.0
pillow~=11.2.1

Flask-Session==0.5.0


spacy>=3.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl